# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Data Ingestion & Privacy Intelligence API",
    "version": "1.0.0",
    "description": "Proteccio Discover backend covering Week 1 ingestion, Week 2 discovery/classification/mapping/profiling, and Week 3 governance/compliance intelligence: privacy risk, compliance controls, remediation workflows, dashboards, reporting, RBAC, alerting, and global search.\n\n**Auth:** When `JWT_SECRET` is set, use `POST /api/auth/login` and send `Authorization: Bearer <JWT>`. When only `API_KEY` is set, send `Authorization: Bearer <API_KEY>` or `X-API-Key: <API_KEY>` on `/api/*` except `/health` and documentation routes.\n\n**Docs UI:** `GET /docs` (Swagger UI). **Raw spec:** `GET /openapi.json`."
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:3000"
        }
      }
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyHeader": []
    }
  ],
  "paths": {
...
    "/api/risk/analyze": {
      "post": {
        "tags": [
          "Risk"
        ],
        "summary": "Full privacy risk analysis from discovery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskAnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "analysis, assessment, profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAnalyzeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/high-risk-datasets": {
      "get": {
        "tags": [
          "Risk"
        ],
        "summary": "High/critical risk datasets from catalog",
        "parameters": [
          {
            "name": "minLevel",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "high",
                "critical"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "high-risk dataset list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/prioritization": {
      "get": {
        "tags": [
          "Risk"
        ],
        "summary": "Ranked remediation queue",
        "parameters": [
          {
            "name": "minLevel",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "medium",
                "high",
                "critical"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "prioritized items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/risk/compliance": {
      "post": {
        "tags": [
          "Risk"
        ],
        "summary": "Compliance intelligence (regulations, controls, flags, remediation)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceIntelligenceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "compliance intelligence report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceIntelligenceResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT from POST /api/auth/login when JWT_SECRET is set; otherwise API_KEY value."
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Same value as API_KEY when set."
      }
    },
    "schemas": {
...
      "ComplianceControlHints": {
        "type": "object",
        "description": "Governance control attestations used by the compliance intelligence engine.",
        "properties": {
          "retentionPolicyIndicated": {
            "type": "boolean"
          },
          "consentManagementIndicated": {
            "type": "boolean"
          },
          "privacyNoticeIndicated": {
            "type": "boolean"
          },
          "lawfulBasisDocumented": {
            "type": "boolean"
          },
          "accessControlsIndicated": {
            "type": "boolean"
          },
          "breachNotificationProcessIndicated": {
            "type": "boolean"
          },
          "dataPrincipalRightsProcessIndicated": {
            "type": "boolean"
          },
          "baaInPlace": {
            "type": "boolean"
          },
          "phiAuditLoggingIndicated": {
            "type": "boolean"
          },
          "optOutMechanismIndicated": {
            "type": "boolean"
          },
          "ismsRiskAssessmentIndicated": {
            "type": "boolean"
          },
          "purposeLimitationDocumented": {
            "type": "boolean"
          },
          "crossBorderSafeguardsIndicated": {
            "type": "boolean"
          },
          "consumerDisclosureIndicated": {
            "type": "boolean"
          },
          "ismsDocumented": {
            "type": "boolean"
          }
        }
      },
      "ExposureHints": {
        "type": "object",
        "properties": {
          "hasApiExposureFlow": {
            "type": "boolean"
          },
          "hasReplicationOrBackupFlow": {
            "type": "boolean"
          },
          "isPubliclyExposed": {
            "type": "boolean"
          },
          "encryptionIndicated": {
            "type": "boolean"
          },
          "crossDatasetDuplicateGroupCount": {
            "type": "integer",
            "minimum": 0
          },
          "unmappedDataset": {
            "type": "boolean"
          },
          "noLineageFlows": {
            "type": "boolean"
          },
          "daysSinceLastActivity": {
            "type": "integer",
            "minimum": 0
          },
          "complianceControls": {
            "$ref": "#/components/schemas/ComplianceControlHints"
          }
        }
      },
      "RiskAnalyzeRequest": {
        "type": "object",
        "required": [
          "discovery"
        ],
        "properties": {
          "discovery": {
            "type": "object",
            "description": "DiscoveryScanResult"
          },
          "classification": {
            "type": "object",
            "description": "Optional ClassificationScanResult"
          },
          "records": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "exposureHints": {
            "$ref": "#/components/schemas/ExposureHints"
          },
          "weights": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "minimum": 0,
              "maximum": 1
            },
            "description": "Optional risk factor weight overrides."
          },
          "profilingOptions": {
            "type": "object"
          }
        },
        "example": {
          "discovery": {
            "trace": {
              "sourceType": "database",
              "sourceName": "prod",
              "entityName": "customers"
            },
            "scannedRecords": 1,
            "findingsPerRecord": [],
            "summary": {
              "aadhaar": 1
            }
          },
          "exposureHints": {
            "encryptionIndicated": false,
            "isPubliclyExposed": true,
            "complianceControls": {
              "retentionPolicyIndicated": false,
              "accessControlsIndicated": false
            }
          }
        }
      },
      "RiskAnalyzeResponse": {
        "type": "object",
        "properties": {
          "analysis": {
            "type": "object"
          },
          "assessment": {
            "type": "object"
          },
          "profile": {
            "type": "object"
          }
        }
      },
      "ComplianceIntelligenceRequest": {
        "type": "object",
        "required": [
          "discovery"
        ],
        "properties": {
          "discovery": {
            "type": "object"
          },
          "classification": {
            "type": "object"
          },
          "exposureHints": {
            "$ref": "#/components/schemas/ExposureHints"
          },
          "factorContributions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "ComplianceIntelligenceResponse": {
        "type": "object",
        "properties": {
          "compliance": {
            "type": "object"
          },
          "regulations": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GDPR",
                "DPDP",
                "HIPAA",
                "CCPA",
                "ISO27001"
              ]
            }
          },
          "controls": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "flags": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "remediationActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
...
  }
}